perf!: reduce margin modal component
remove n/a from withdraw health
